Move datapath into UART module
The next commit will introduce a USB datapath that's API compatible
with the UART datapath...
